Allow explicit aux sets in chor* and forbid self-reference without one
This restriction is necessary because we currently don't perform any sort of
smart constraint-based inference. Specifically, it could happen that a chor*
whose aux set we want to infer uses its own name to pass itself to some other
choreography, at which point we would need the aux to type check the call, but
don't have it because we haven't analyzed the body yet.

Calls to `recur` are not a problem because they don't meaningfully modify the
aux set of a chor*. We just have to take care to update their mentions
after-the-fact, once they have been computed by analyzing the body.